requirements.txt

streamlit
pandas
gspread
google-auth
transformers
plotly
fpdf
